feat: add file-based JSON mapping for Todo detail and implement domain model compatibility tests
