+++ examples/prisma-comparison-demo/rawsql-ts/searchTodos.json
@@ -1,0 +1,43 @@
+{
+    "rootName": "todos",
+    "rootEntity": {
+        "id": "t",
+        "name": "todo",
+        "columns": {
+            "todoId": "todo_id",
+            "title": "title",
+            "description": "description",
+            "completed": "completed",
+            "createdAt": "created_at",
+            "updatedAt": "updated_at",
+            "commentCount": "comment_count"
+        }
+    },
+    "nestedEntities": [
+        {
+            "id": "u",
+            "name": "user",
+            "parentId": "t",
+            "propertyName": "user",
+            "relationshipType": "object",
+            "columns": {
+                "userId": "user_id",
+                "userName": "user_name",
+                "email": "email"
+            }
+        },
+        {
+            "id": "c",
+            "name": "category",
+            "parentId": "t",
+            "propertyName": "category",
+            "relationshipType": "object",
+            "columns": {
+                "categoryId": "category_id",
+                "categoryName": "category_name",
+                "color": "color"
+            }
+        }
+    ],
+    "resultFormat": "array"
+}
